requirements.txt
flask
twilio
pillow
pytesseract
gspread
oauth2client
requests
